ports: lua

+++ util/ports.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
+# IMPORTANT: run from the root of the repo!
 
-# The cwd has to be ports/.. for these to work
 ./ports/bash-prebuilt.sh
 ./ports/busybox.sh
 ./ports/doomgeneric.sh
@@ -8,7 +8,12 @@
 ./ports/neofetch.sh
 ./ports/ncurses.sh
 ./ports/vim.sh
+./ports/tree.sh
+./ports/lua.sh
 
-echo "Cleaning up..."
-rm -rf ports/src/
-cd ..
+read -p "Clean up? (Y/n) " answer
+answer=${answer:-Y}
+if [[ "$answer" =~ ^[Yy]$ ]]; then
+    echo "Cleaning up..."
+    rm -rf ports/src/
+fi
